rest/requirements.txt
uvicorn
fastapi
sqlmodel
alembic
pymysql
cryptography
python-dotenv
s3fs
websockets
Pillow
PyMongo